QR-Code page
This is a page page where pet owners can create a QR-code that can be used to attched it to their pet, use it in flyers and to share it.  With this QR-code a pet finder can easly contact the owner of the pet.

The page should a QR-code generator, where the pet owner can fill in their  contact details.
It should have a share button to share the QR-code.
It should a print button to be able to print the QR-code.
There also should be an explanation what poeple can do with the QR-code.
The main menu bar schould be placed at the bottom of the page.
 
Please look at the attchment for the proptotype page.


TO DO:

still need to add the bottom menu.